// src/app/interfaces/user.ts
export interface User {
    name?: string;
    email: string;
    phone?: string;
    nickName: string;
    login: string;
    uid: string;
    provider: string;
    photoURL: string;
    accessToken: string;
    profileImage: string;
    phoneNumber?: string;
    verificationId?: string;
    displayName?: string;
    online: boolean;
  }

  